Refactor register command back into cli
